+++ stories/templates/collapser/default.template.js
@@ -1,5 +1,5 @@
 export const template = `
-<div>
+<div class="storybook-components--collapser">
   <h4>Change window size to check up how component works.</h4>
   <st-collapser :elements="elements"></st-collapser>
 </div>